Fixing test case

+++ modules/swagger-parser-v2-converter/src/test/java/io/swagger/parser/test/V2ConverterTest.java
@@ -164,7 +164,7 @@ public class V2ConverterTest {
     private static final int MIN_LENGTH = 3;
     private static final int NUMBER_VALUE_TWENTY = 20;
     private static final double MULTIPLE_OF_VALUE = 0.01D;
-    private static final long DEFAULT_VALUE = 11L;
+    private static final int DEFAULT_VALUE = 11;
     private static final int EXAMPLE_8_NUMBER = 8;
     private static final int EXAMPLE_42_NUMBER = 42;
